Allow wider range of config changes.
This change adds a dictionary that represents the previous defaults for the `RaygunSender`.  Those defaults are now merged with the incoming config dict that's passed into `RaygunSender.__init__()`.  If no config value is present, we just use the defaults, otherwise, any of the previously hardcoded values are now overridable by passing in the appropriate config option.

The big API decision here that I made was to represent the camelCase attributes of `RaygunSender` as snake_case keys in the config dictionary.  That was a conscious choice to make the configuration and usage of this a little more PEP8-ish, without actually altering any of the sending code itself.  So there's a transformation step that happens with the calls to `normalize_dict` from `RaygunSender.__init__()`