Fix RawPostDataException on request.body access.
Django's HttpRequest class doesn't like the `request.body` to be accessed more than one time.  Upon the second attempt to read from `request.body`, Django throws a `RawPostDataException`.  Since the previous code in this spot was conditional upon a `hasattr(request, 'body')` check, and because hasattr returns False when the attribute access raises an exception, raygun was falling back to looking up the rawData field from `request.raw_post_data`.  But that attribute doesn't exist in newer versions of Django, so we were getting exceptions in the raygun4py's Django process_exception middleware, which is a royal pain in the tush.

So, this change does the hasattr check on both attributes, and only includes rawData if one of the `request.body` or `request.raw_post_data` attributes actually exists and is readable.