Fix #23: Use repr on localvars instead of str.
When converting localvars into a readable format, str() fails to decode
unicode strings (as opposed to strings with unicode bytes, which is what
is currently tested). unicode() will also fail without being told an
encoding, and we can't assume that everything is utf8. repr() works for
all values and seems like a better choice for these values anyway.